Add `cmd` command
`cmd` should used to extend the bot.
It is not planned to implement many commands in the bot but to point to
a specific directory and have executables/scripts in there.
The bot will call those executables via `!cmd`.

By default the the directory is the current working directory.
If one wants to change this one should use the `-d` option.